backend/pyproject.toml
[project]
name = "game-backend"
version = "0.1.0"
description = "FastAPI backend for multiplayer game — player accounts and save/load"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "sqlalchemy>=2.0.0",
    "pydantic>=2.0.0",
    "bcrypt>=4.0.0",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.18",
    "websockets>=14.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.25.0",
    "coverage>=7.0.0",
    "ruff>=0.11.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
target-version = "py313"
line-length = 120
lint.select = ["E", "F", "W", "I", "UP"]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.coverage.run]
source = ["app"]
omit = ["tests/*"]

[tool.coverage.report]
show_missing = true
fail_under = 70
